@@ -2,6 +2,8 @@
 #define __APPLICATION_H__
 
 #include <SDL.h>
+#include "config.h"
+#include "raster.h"
 
 class Application
 {
@@ -13,53 +15,57 @@ public:
     virtual void run(Application* the_app)
     {
         bool running = true;
-        app = the_app;
+        _app = the_app;
 
 		SDL_Init(SDL_INIT_VIDEO);
 
         init();
 
-		gWindow = SDL_CreateWindow(info.title, 100, 100, info.windowWidth, info.windowHeight, SDL_WINDOW_SHOWN);
+		_gWindow = SDL_CreateWindow(_info.title, 100, 100, _info.winWidth, _info.winHeight, SDL_WINDOW_SHOWN);
 
-		gRender = SDL_CreateRenderer(gWindow, -1, SDL_RENDERER_ACCELERATED|SDL_RENDERER_PRESENTVSYNC);
+		_gRender = SDL_CreateRenderer(_gWindow, -1, SDL_RENDERER_ACCELERATED|SDL_RENDERER_PRESENTVSYNC);
 
 		//创建texture,注意像素格式和访问方式
-		gTexture = SDL_CreateTexture(gRender, SDL_PIXELFORMAT_RGBA8888, SDL_TEXTUREACCESS_STREAMING, info.windowWidth, info.windowHeight);
+		_gTexture = SDL_CreateTexture(_gRender, SDL_PIXELFORMAT_RGBA8888, SDL_TEXTUREACCESS_STREAMING, _info.winWidth, _info.winHeight);
 
-		format = SDL_AllocFormat(SDL_PIXELFORMAT_RGBA8888);
+		_gFormat = SDL_AllocFormat(SDL_PIXELFORMAT_RGBA8888);
+
+		_gRaster.setInfo(_info.winWidth, _info.winHeight, _gFormat);
 
         startup();
 
 		void* pix;
-		int pitch;
+		int32 pitch;
         do
         {
-			SDL_PollEvent(&event);
-			if(event.type == SDL_QUIT)
+			SDL_PollEvent(&_event);
+			if(_event.type == SDL_QUIT)
 				break;
-			SDL_LockTexture(gTexture, NULL, &pix, &pitch);
-			render(pix, pitch);
-			SDL_UnlockTexture(gTexture);
+			SDL_LockTexture(_gTexture, NULL, &pix, &pitch);
+			_gRaster.setBuffer(pix);
+			_gRaster.clear();
+			render();
+			SDL_UnlockTexture(_gTexture);
 
-			SDL_RenderCopy(gRender, gTexture, NULL, NULL);
-			SDL_RenderPresent(gRender);
+			SDL_RenderCopy(_gRender, _gTexture, NULL, NULL);
+			SDL_RenderPresent(_gRender);
         } while (running);
 
         shutdown();
 
-		SDL_FreeFormat(format);
-		SDL_DestroyRenderer(gRender);
-		SDL_DestroyTexture(gTexture);
-		SDL_DestroyWindow(gWindow);
+		SDL_FreeFormat(_gFormat);
+		SDL_DestroyRenderer(_gRender);
+		SDL_DestroyTexture(_gTexture);
+		SDL_DestroyWindow(_gWindow);
 		SDL_Quit();
     }
 
     virtual void init()
     {
-        strcpy(info.title, "RenderEngine");
+        strcpy(_info.title, "RenderEngine");
 
-        info.windowWidth = 800;
-        info.windowHeight = 600;
+        _info.winWidth = 800;
+        _info.winHeight = 600;
     }
 
     virtual void startup()
@@ -67,7 +73,7 @@ public:
 
     }
 
-	virtual void render(void* pixel, int pitch) 
+	virtual void render() 
     {
 
     }
@@ -81,19 +87,21 @@ public:
     struct APPINFO
     {
         char title[128];
-        int windowWidth;
-        int windowHeight;
+        int32 winWidth;
+        int32 winHeight;
     };
 
 protected:
-    APPINFO     info;
-    static      Application *app;
+    APPINFO     _info;
+    static      Application *_app;
 
-	SDL_Window* gWindow = NULL;
-	SDL_Renderer* gRender = NULL;
-	SDL_Texture* gTexture = NULL;
-	SDL_Event event;
-	SDL_PixelFormat* format;
+	SDL_Window* _gWindow = nullptr;
+	SDL_Renderer* _gRender = nullptr;
+	SDL_Texture* _gTexture = nullptr;
+	SDL_Event _event;
+	SDL_PixelFormat* _gFormat = nullptr;
+
+	render::Raster _gRaster;	
 };
 
 #define DECLARE_MAIN(a)                             \
@@ -104,6 +112,6 @@ int main(int argc, const char ** argv)              \
     delete app;                                     \
     return 0;                                       \
 }
-Application * Application::app = 0;
+Application * Application::_app = nullptr;
 #endif /* __APP_H__ */
 
