@@ -1,14 +1,15 @@
 /*
  *	basic type
  */
-// typedef signed char             int8;
-// typedef unsigned char           uint8;
-// typedef signed short            int16;
-// typedef unsigned short          uint16;
-// typedef signed int              int32;
-// typedef unsigned int            uint32;
-// typedef signed long long int    int64;
-// typedef unsigned long long int  uint64;
+#include <stdint.h>
+typedef int8_t 					int8;
+typedef uint8_t 				uint8;
+typedef int16_t 				int16;
+typedef uint16_t 				uint16;
+typedef int32_t 				int32;
+typedef uint32_t 				uint32;
+typedef int64_t 				int64;
+typedef uint64_t 				uint64;
 // typedef float                   float32;
 // typedef double                  float64;
 // #ifdef _64BIT
@@ -26,5 +27,3 @@
 /*
  * app define
  */
-#define WIN_WIDTH 640
-#define WIN_HEIGHT 480
