@@ -8,10 +8,10 @@ namespace render {
 class Rgba4Byte {
 public:
 	Rgba4Byte(
-			uint8_t r = 255,
-			uint8_t g = 255,
-			uint8_t b = 255,
-			uint8_t a = 255
+			uint8 r = 255,
+			uint8 g = 255,
+			uint8 b = 255,
+			uint8 a = 255
 			)
 	{
 		_r = r;
@@ -19,7 +19,7 @@ public:
 		_b = b;
 		_a = a;
 	}
-	friend bool operator = (const Rgba4Byte& left, const Rgba4Byte& right) 
+	friend bool operator == (const Rgba4Byte& left, const Rgba4Byte& right) 
 	{
 		return left._r == right._r
 			&& left._g == right._g
@@ -40,10 +40,10 @@ public:
 		return (_b) | (_g << 8) | (_r << 16) | (_a << 24);
 	}
 public:
-	uint8_t _r;
-	uint8_t _g;
-	uint8_t _b;
-	uint8_t _a;
+	uint8 _r;
+	uint8 _g;
+	uint8 _b;
+	uint8 _a;
 };
 typedef Rgba4Byte Rgba;
 }
