@@ -1,52 +1,34 @@
 #ifndef RASTER_H
 #define RASTER_H
 
+#include <SDL.h>
+#include "config.h"
 #include "math.h"
 
 namespace render {
 
+typedef Uint32 BufferType;
+
 class Raster {
 public:
-	Rgba _buffer[256][256];
+	int32 _width = 0;
+	int32 _height = 0;
+private:
+	BufferType* _buffer = nullptr;
+	SDL_PixelFormat* _colorFormat = nullptr;
 public:
-	void clear() 
+	void clear();
+	void drawPoint(int32 x, int32 y, Rgba color, int32 ptSize);
+public:
+	inline void setInfo(int32 width, int32 height, SDL_PixelFormat* colorFormat) { _width = width; _height = height; _colorFormat = colorFormat; }
+	inline void setBuffer(void* buffer) { _buffer = (BufferType*)buffer;}
+	inline void setPixel(uint32 x, uint32 y, Rgba color) 
 	{
-		memset(&_buffer, 0, sizeof(_buffer));
-	}
-	void setPixel(int32_t x, int32_t y, const Rgba& color) 
-	{
-		if ( x < 0 || y < 0 || x >= 256 || y >= 256 )
+		if ( x >= _width || y >= _height )
 			return;
-		_buffer[y][x] = color;
-	}
-	void drawPoint(int32_t x, int32_t y, const Rgba& color, int32_t ptSize) 
-	{
-		switch(ptSize) 
-		{
-			case 1:
-				setPixel(x, y, color);
-				break;
-			case 2:
-				setPixel(x + 0, y + 0, color);
-				setPixel(x + 1, y + 0, color);
-				setPixel(x + 0, y + 1, color);
-				setPixel(x + 1, y + 1, color);
-				break;
-			case 3:
-				setPixel(x - 1, y - 1, color);
-				setPixel(x + 0, y - 1, color);
-				setPixel(x + 1, y - 1, color);
-
-				setPixel(x - 1, y + 0, color);
-				setPixel(x + 0, y + 0, color);
-				setPixel(x + 1, y + 0, color);
-
-				setPixel(x - 1, y + 1, color);
-				setPixel(x + 0, y + 1, color);
-				setPixel(x + 1, y + 1, color);
-				break;
-		}
+		_buffer[y * _width + x] = SDL_MapRGBA(_colorFormat, color._r, color._g, color._b, color._a);
 	}
 };
+
 }
 #endif
